requirements.txt
altair==4.0.1
argon2-cffi==21.1.0
astor==0.8.1
base58==2.1.1
beautifulsoup4==4.12.2
bleach==3.3.1
blinker==1.5
boto3==1.16.63
botocore==1.19.63
bottleneck==1.3.7
cachetools==4.2.4
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.5
entrypoints==0.3
enum-compat==0.0.3
faker==18.9.0
favicon==0.7.0
htbuilder==0.6.1
i==2.1.0
idna==2.10
# importlib-metadata==2.1.3
importlib-metadata
ipykernel==5.5.6
ipython==7.9.0
ipywidgets==7.7.5
jedi==0.17.2
jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.6.3
lightgbm==3.3.0
lxml==4.9.2
markdown==3.4.3
markdownlit==0.0.7
markupsafe==1.1.1
more-itertools==9.1.0
nbconvert==5.6.1
notebook==6.1.6
numexpr==2.8.4
numpy==1.21.6
openpyxl==3.1.2
# packaging==20.9
packaging
# pandas==2.0.0
pandas==1.3.2
parso==0.7.1
pathtools==0.1.2
patsy==0.5.3
PDPbox==0.2.0
# pillow==7.2.0
pillow
pip==20.3.4
plotly==5.14.1
plotly-express==0.4.0
prometheus-client==0.12.0
prompt-toolkit==2.0.10
protobuf==3.19.6
pydeck==0.6.2
pymdown-extensions==10.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2023.3
pywin32==301
pywinpty==0.5.7
pyzmq==20.0.0
requests==2.25.1
s3transfer==0.3.7
scipy==1.10.1
soupsieve==2.4.1
st-annotated-text==4.0.0
statsmodels==0.13.5
streamlit==1.21.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
terminado==0.8.3
traitlets==4.3.3
tzdata==2023.3
urllib3==1.26.9
validators==0.20.0
# watchdog==0.10.7
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==3.6.4
win-unicode-console==0.5
zipp==1.2.0
